deps.windows: Re-enable mbedtls
This enables mbedtls to be compiled natively on msvc.
The lib is built as a static dependency because mbedtls dlls can not
be built natively with msvc due to long-standing bugs.

Note: for future upgrades, 3.3.0 has a bug with designated initializers
which prevents the building of libdatachannel on msvc. But 3.4.0 does
not.

Signed-off-by: pkv <pkv@obsproject.com>