deps.ffmpeg: Don't build libdatachannel on windows
Since libdatachannel is now built natively on msvc, this updates the
script in deps.ffmpeg so that it does not cross-compile the lib anymore.
In addition, this updates libdatachannel to the commit supporting
simulcast and removes the now unnecessary libdatachannel patch which is
in the tree which is checked out.

Signed-off-by: pkv <pkv@obsproject.com>